+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 (unreleased)
 
 - Added column alias
+- Raise `ActiveRecord::UnknownAttributeReference` for non-attribute arguments
 - Dropped support for Active Record < 5.2
 
 ## 0.2.8 (2021-01-16)